Frontend: _ Correciones en las queries y mutations - Añadidas nuevas queries y mutations Backend: - La tabla votos admite decimales y no solo enteros.

+++ Backend/database/migrations/2026_02_27_111300_create_votes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('votes', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->enum('type', ['vote', 'skip'])->default('vote');
-            $table->integer('score')->nullable();
+            $table->decimal('score', 4, 1)->nullable();
 
             $table->foreignUuid('user_id')->constrained('users')->cascadeOnDelete();
             $table->foreignUuid('item_id')->constrained('items')->cascadeOnDelete();
